requirements.txt
aiofiles==0.5.0
aioredis==2.0.0
aiosmtplib==1.1.6
aniso8601==7.0.0
anyio==3.4.0
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.1
bcrypt==3.2.0
black==21.8b0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-mail==1.0.2
fastapi-pagination==0.9.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
itsdangerous==1.1.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
orjson==3.6.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.3.0
promise==2.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
pyodbc==4.0.32
pyparsing==3.0.6
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==4.0.2
regex==2021.8.28
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.23
starlette==0.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
ujson==4.1.0
urllib3==1.26.6
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
wrapt==1.13.3
